New Project with Nav Drawer + Swipe View
I made a new project so that I don't accidentally wreck the old one.
This project includes the navigation drawer that is supposed to be
present throughout the whole app and fragments within the input
activity.

Right now, there are two errors: 1. R.layout is no longer recognized,
but that's an easy fix 2. PagerAdapter and getSupportFragmentManager are
not compatible for some reason